Moving files to buildpack instead of venia-concept.
