+++ packages/pwa-buildpack/lib/Utilities/fetcherUtils.js
