+++ packages/pwa-buildpack/lib/Utilities/index.js
@@ -2,5 +2,6 @@ module.exports = {
     addImgOptMiddleware: require('./addImgOptMiddleware'),
     configureHost: require('./configureHost'),
     createDotEnvFile: require('./createDotEnvFile'),
-    loadEnvironment: require('./loadEnvironment')
+    loadEnvironment: require('./loadEnvironment'),
+    fetcherUtils: require('./fetcherUtils')
 };
