+++ packages/pwa-buildpack/lib/WebpackTools/index.js
@@ -4,5 +4,6 @@ module.exports = {
     ServiceWorkerPlugin: require('./plugins/ServiceWorkerPlugin'),
     MagentoResolver: require('./MagentoResolver'),
     PWADevServer: require('./PWADevServer'),
-    UpwardDevServerPlugin: require('./plugins/UpwardDevServerPlugin')
+    UpwardDevServerPlugin: require('./plugins/UpwardDevServerPlugin'),
+    AsyncWebpackPlugin: require('./plugins/AsyncWebpackPlugin')
 };
