+++ packages/pwa-buildpack/scripts/fetch-mock.js
